refactor: remove checks mock adapter, fix warnings, update tests
