+++ tap_core/src/tap_receipt/mod.rs
@@ -4,7 +4,7 @@
 mod receipt;
 mod receipt_auditor;
 mod received_receipt;
-use std::{collections::HashMap, sync::{Arc, RwLock}};
+use std::collections::HashMap;
 
 use alloy_primitives::Address;
 pub use receipt::Receipt;
@@ -15,7 +15,6 @@ pub use received_receipt::{
 };
 
 use serde::{Deserialize, Serialize};
-use strum_macros::{Display, EnumString};
 use thiserror::Error;
 
 use crate::checks::{CheckingChecks, ReceiptCheck};
@@ -72,6 +71,3 @@ pub fn get_full_list_of_checks() -> Vec<ReceiptCheck> {
         // ReceiptCheck::CheckSignature,
     ]
 }
-
-#[cfg(test)]
-pub mod tests;
