--- tap_core/src/tap_receipt/tests/mod.rs
@@ -1,5 +1,0 @@
-// Copyright 2023-, Semiotic AI, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod receipt_tests;
-pub mod received_receipt_tests;
